# openapi.yaml
openapi: 3.1.1
info:
  title: PeerCast API
  # Kept equal to the root package.json SSOT by `npm run check:versions`.
  version: 1.1.10
  description: |
    Self-hosted, WebRTC peer-to-peer live broadcast platform. Everything runs on
    Node.js: REST API, PeerJS signaling, SQLite storage, and the installable
    React PWA — served from one process.

    All paths below are mounted under the `/api` base path. Request and response
    bodies are JSON (`Content-Type: application/json`) except the recording chunk
    upload (`PUT /api/recordings/:id/chunk`), which is raw `application/octet-stream`.

    Timestamps are Unix epoch **milliseconds**. Broadcast/report IDs are nanoid
    strings; recording IDs are 18-char hex strings.

    ## Authentication

    - `Authorization: Bearer <accessToken>` on authenticated endpoints.
    - The **refresh token** travels only as an HttpOnly cookie (`rt`, `Path=/api/auth`).
    - `POST /api/auth/refresh` rotates the refresh token and returns a new access token.
    - A short-lived opaque signaling ticket (`POST /api/auth/signaling-ticket`) replaces
      the JWT in the PeerJS WebSocket URL (`?token=`).

    ## Error envelope

    Every error is a JSON object: `{ "error": string, "code"?: string }`.
    Common machine-readable codes: `USERNAME_TAKEN`, `EMAIL_TAKEN`,
    `REGISTRATION_CLOSED`, `INVITE_REQUIRED`, `INVITE_INVALID`, `BAD_CREDENTIALS`,
    `BANNED`, `EMAIL_UNVERIFIED`, `BAD_PASSWORD`, `EMAIL_SAME`, `TOKEN_INVALID`,
...

...

...
servers:
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new account
      description: |
        Rate-limited (5/h per IP). When registration mode is `invite`, an
        `inviteCode` is required. A configured admin handle may pair with the
        operator secret via the `adminBootstrapSecret` body field or the
        `X-Admin-Bootstrap-Secret` header. Sets the `rt` refresh cookie.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created.
          headers:
            Set-Cookie:
              description: HttpOnly refresh-token cookie `rt`.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/Error400"
        "403":
          description: |
            `REGISTRATION_CLOSED`, `INVITE_REQUIRED`, or `INVITE_INVALID` —
            registration gated by the configured mode.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: "`USERNAME_TAKEN` or `EMAIL_TAKEN`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/Error429"
  /auth/login:
    post:
      tags: [Auth]
      summary: Sign in
      description: |
        Rate-limited (10/15 min per IP, 10/15 min per username on failed login
        attempts). Sets the `rt` refresh cookie.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Signed in.
          headers:
            Set-Cookie:
              description: HttpOnly refresh-token cookie `rt`.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: "`BAD_CREDENTIALS`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: "`BANNED` or `EMAIL_UNVERIFIED` (when email verification is required)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/Error429"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: |
        Reads and rotates the HttpOnly `rt` cookie (no request body).
        Rate-limited (30/15 min per IP).
      operationId: refresh
      security:
        - cookieAuth: []
      responses:
        "200":
          description: New access token issued; refresh cookie rotated.
          headers:
            Set-Cookie:
              description: Rotated HttpOnly refresh-token cookie `rt`.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          description: Missing or invalid refresh token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: "`BANNED`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/Error429"
...
  /auth/send-verification:
    post:
      tags: [Email]
      summary: Send (or resend) the verification email
      description: |
        Fire-and-forget; always `{ok:true}` unless unauthenticated. Silent for
        missing accounts. Rate-limited (5/h per IP).
      operationId: sendVerification
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Email queued (or account already verified).
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
                    const: true
                  alreadyVerified:
                    type: boolean
                    description: True when the account was already verified.
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "429":
          $ref: "#/components/responses/Error429"
  /auth/verify-email:
    get:
      tags: [Email]
      summary: Verify an email address from a link
      description: |
        Consumes the single-use 24 h token and 302-redirects to the web app.
        Invalid/expired tokens redirect with `?status=invalid`. Rate-limited
        (30/h per IP).
      operationId: verifyEmail
      parameters:
        - name: token
          in: query
          required: true
          description: Email verification token from the confirmation link.
          schema:
            type: string
      responses:
        "302":
          description: |
            Redirects to `${appUrl}/verify-email?status=ok` on success or
            `?status=invalid` otherwise.
          headers:
            Location:
              schema:
                type: string
        "400":
          description: Missing `token` parameter.
  /auth/change-email:
    post:
      tags: [Email]
      summary: Request an email change (verified, two-step)
      description: |
        Re-authenticates with the current password, then emails a confirmation
        link to the new address. The new email is NOT applied until the link is
        clicked. Rate-limited (5/h per IP, 3/h per target address).
      operationId: changeEmail
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeEmailRequest"
      responses:
        "200":
          $ref: "#/components/responses/OkResponse"
        "400":
          description: "Validation, `BAD_PASSWORD`, or `EMAIL_SAME`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "409":
          description: "`EMAIL_TAKEN`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/Error429"
...
  /auth/forgot-password:
    post:
      tags: [Email]
      summary: Send a password-reset email
      description: |
        Always returns 200 (no account enumeration). Sends a reset link when the
        address exists. Rate-limited (5/h per IP, 3/h per address).
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          $ref: "#/components/responses/OkResponse"
        "429":
          $ref: "#/components/responses/Error429"
  /auth/reset-password:
    post:
      tags: [Email]
      summary: Consume a reset token and set a new password
      description: |
        Token is single-use with a 1 h TTL; the row is deleted on any consume
        attempt, so replay always fails. Revokes all refresh tokens.
        Rate-limited (10/h per IP).
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          $ref: "#/components/responses/OkResponse"
        "400":
          description: "`TOKEN_INVALID` or validation."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/Error429"

  # ─── Users ───────────────────────────────────────────────────────────────
...
  /users/{username}/follow:
    post:
      tags: [Users]
      summary: Follow a user
      description: Idempotent. Rate-limited (60/15 min per IP).
      operationId: follow
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UsernamePath"
      responses:
        "200":
          description: Now following.
          content:
            application/json:
              schema:
                type: object
                required: [ok, following]
                properties:
                  ok:
                    type: boolean
                    const: true
                  following:
                    type: boolean
                    const: true
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
        "429":
          $ref: "#/components/responses/Error429"
    delete:
      tags: [Users]
      summary: Unfollow a user
      description: Idempotent.
      operationId: unfollow
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UsernamePath"
      responses:
        "200":
          description: No longer following.
          content:
            application/json:
              schema:
                type: object
                required: [ok, following]
                properties:
                  ok:
                    type: boolean
                    const: true
                  following:
                    type: boolean
                    const: false
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"

  # ─── Broadcasts ──────────────────────────────────────────────────────────
...
  /broadcasts:
    post:
      tags: [Broadcasts]
      summary: Start a broadcast
      description: |
        The host picks a fresh random `peerId` and answers PeerJS calls on it.
        `accessCode` is required when `access` is `code` and is stored only as a
        SHA-256 digest. Starting ends any prior live broadcast for this user;
        public broadcasts notify opted-in followers (fire-and-forget).
      operationId: startBroadcast
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartBroadcastRequest"
      responses:
        "201":
          description: Broadcast started. `peerId` is nulled in this payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartBroadcastResponse"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
...
  /reports:
    post:
      tags: [Reports]
      summary: File an abuse report
      description: |
        Auth optional; the reporter handle is recorded when signed in.
        Rate-limited (10/h per IP).
      operationId: createReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportRequest"
      responses:
        "201":
          description: Report created.
          content:
            application/json:
              schema:
                type: object
                required: [ok, report]
                properties:
                  ok:
                    type: boolean
                    const: true
                  report:
                    $ref: "#/components/schemas/Report"
        "400":
          description: Validation failure or reporting yourself.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/Error404"
        "429":
          $ref: "#/components/responses/Error429"

  # ─── Admin ───────────────────────────────────────────────────────────────
  # NOTE: every /admin path requires an authenticated admin (bearerAuth).
  # Non-admins get 403 `ADMIN_ONLY`; anonymous callers get 401.
  /admin/users:
    get:
      tags: [Admin]
      summary: List users (admin)
      description: Newest first, limit 100. Optional filter across username/display name/email. Includes banned.
      operationId: adminListUsers
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          description: Text filter.
          schema:
            type: string
      responses:
        "200":
          description: Admin users.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          description: "`ADMIN_ONLY`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /admin/reports:
    get:
      tags: [Admin]
      summary: List reports (admin)
      description: Newest first, limit 200. Invalid `status` falls back to `open`.
      operationId: adminListReports
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          description: Report status filter.
          schema:
            type: string
            enum: [open, resolved, dismissed, all]
            default: open
      responses:
        "200":
          description: Reports + open count.
          content:
            application/json:
              schema:
                type: object
                required: [reports, openCount]
                properties:
                  reports:
                    type: array
                    items:
                      $ref: "#/components/schemas/Report"
                  openCount:
                    type: integer
                    description: Number of currently open reports.
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          description: "`ADMIN_ONLY`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /push/subscribe:
    post:
      tags: [Push]
      summary: Subscribe to push notifications
      description: |
        Endpoint host must be a known push service. Upserts the subscription.
        Rate-limited (5/h per IP).
      operationId: pushSubscribe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscribeRequest"
      responses:
        "200":
          $ref: "#/components/responses/OkResponse"
        "400":
          description: "Push disabled, `PUSH_ENDPOINT`, or validation."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "429":
          $ref: "#/components/responses/Error429"
    delete:
      tags: [Push]
      summary: Unsubscribe from push notifications
      description: The endpoint record is removed only if owned by the caller.
      operationId: pushUnsubscribe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushUnsubscribeRequest"
      responses:
        "200":
          $ref: "#/components/responses/OkResponse"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"

  # ─── Recordings ──────────────────────────────────────────────────────────
  # NOTE: every /recordings path requires auth and is scoped to the caller's
  # own recordings (owner-only reads/writes/replay/delete).
  /recordings:
    get:
      tags: [Recordings]
      summary: My recordings
      description: Newest first.
      operationId: listRecordings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Own recordings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingListResponse"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
    post:
      tags: [Recordings]
      summary: Start a recording
      description: |
        Creates the recording row (status `recording`, `sizeBytes`/`durationMs`
        zero). `broadcastId`, when given, must belong to the caller.
        Rate-limited (20/h per IP).
      operationId: createRecording
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecordingRequest"
      responses:
        "201":
          description: Recording created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordingResponse"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "429":
          $ref: "#/components/responses/Error429"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token returned by register/login/refresh.
    cookieAuth:
      type: apiKey
      in: cookie
      name: rt
      description: HttpOnly, SameSite refresh-token cookie (Path=/api/auth).
...
  responses:
...
    Error400:
      description: |
        Bad request. Codes include `VALIDATION` (zod), `BAD_JSON` (malformed
        body), `BAD_PASSWORD`, `EMAIL_SAME`, `PUSH_ENDPOINT`, `TOKEN_INVALID`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    RegisterRequest:
      type: object
      required: [username, email, displayName, password]
      properties:
        username:
          type: string
          pattern: ^[a-zA-Z0-9_-]{3,30}$
        email:
          type: string
          format: email
          maxLength: 254
        displayName:
          type: string
          maxLength: 80
        password:
          type: string
          format: password
          minLength: 8
        about:
          type: [string, "null"]
          maxLength: 500
        inviteCode:
          type: string
          maxLength: 64
          description: Required when the server's registration mode is `invite`.
        adminBootstrapSecret:
          type: string
          maxLength: 256
          description: |
            Required only when bootstrapping a configured administrator handle.
            May also be sent as the `X-Admin-Bootstrap-Secret` header.
...